// BackEnd/src/types/transactions.type.ts
import type { TransactionType } from '@prisma/client';
import type { categorySummary } from './category.types';

export interface TransactionFilter {
  userId: string;
  date?: {
    gte: Date;
    lte: Date;
  };
  type?: TransactionType;
  categoryId?: string;
}

export interface TransactionSummary {
  totalExpenses: number;
  totalIncomes: number;
  balance: number;
  expensesByCategory: categorySummary[];
}